overhauling the CMake architecture to make it easier to manage
The main difference is that the code MD code is compiled as a static library.
This has the advantage that dependencies on MPI and Eigen are automatically
propagated to other targets in the repository.

I've removed the `cmake/` directory. CMake's FetchContent package now supersedes
`tools.cmake`, and makes `Eigen3.cmake` obsolete since `find_package()` can
automatically work in tandem with FetchContent.

I've removed the root `main.cpp` file, and instead added a `milestones/`
directory, where the students are expected to put their production simulations.

Finally, I've added detailed instructions on how to add files in the various
directories, and populated the `.gitignore` file with a large spectrum of
patterns to avoid accidentally committing unwanted files.